server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pillow==10.0.0
numpy>=1.20.0
opencv-python==4.8.0.76
torch>=2.0.0
torchvision>=0.15.0
pycocotools>=2.0.6 